requirements.txt
langchain-openai
pandas
numpy
python-dotenv
langchain_community
langchain_core
streamlit